requirements.txt
# Core dependencies
flask==2.3.3
requests==2.31.0
googlemaps==4.10.0
flask-cors==4.0.0
python-dotenv==1.0.0
